More fixes (#239)
* Add accent to focused button

* Add slowdown during spell aim

* Increase slow

* Update spell tutorial message

* Restructure folder heirarchy and move blacksmith quest trigger

* Create HUD highlighter

* Add UI highlighter to quest completion

* Increase radius of rotate tutorial

* Only show trail on new quests

* Fix leave Glade Village starting early

* Fix alchemist saved ink dialogue

* TEMP FIX: Rabbit slime transition to dodge

There's an issue caused by SteeringAgents reached goal flag making the
velocity to zero which in turn causes isMoving to be set to zero.

This forces a transition back to idle. Problem is this takes precedence
over a transition to dodge. I've disabled the transition back to idle
which should look fine for now as the Rabbit Slime is rarely ever
idling.

* Cut rabbit animation to jump earlier

* Adjust dialogue